refactor: use right db account permission for right query; refactor some code into pattern
