+++ src/main/java/com/amkor/common/utils/SharedConstValue.java
@@ -5,4 +5,5 @@ public class SharedConstValue {
     public static final int FACTORY_ID = 80;
     public static final int SITE_ID = 1;
     public static final String PLANT = "V1";
+    public static final String AMKOR_SHORTNAME = "ATV";
 }
